Special case Gamma distribution for limit at zero
Strictly speaking, zero is not part of the domain of the Gamma distribution.
However, as we are doing floating point arithmetic small values may get rounded
to zero, and so for these values it would be good to produce a sensible result
(even if positive or negative infinity), rather than failing due to a floating
point exception caused by computing the logarithm of zero.

Later, it may be good to undue this special casing and instead avoid that the
function is called for out-of-domain values.