Avoid trying to normalize rates when count is zero
In all likelihood, after a few iterations rates for genes in spots which
have zero counts for that gene will reach 0. Then division of the rates
by the summed rate will result in a division by zero, and hence program
abortion. This commit fixes it by only trying to perform this
normalization when counts are nonzero.